pyproject.toml
[project]
name = "adk-yt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.13.0",
    "google-generativeai>=0.8.5",
]
